GPU CUDA 12/requirements.txt
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

pillow
tqdm
numpy
jsonschema
